Move Promise to Deferred converter into deferred utils module
